feat(*): move core config to distinct package
