+++ bot/core/models/boil_types.go
