+++ bot/core/models/core_config.go
